Fixing things

+++ CallCenterService/Models/Repair.cs
@@ -17,7 +17,7 @@ namespace CallCenterService.Models
         public String Description { get; set; }
         [Required]
         public DateTime  Date { get; set; }
-        public float Price { get; set; }
-        public float PartsPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal PartsPrice { get; set; }
     }
 }
